Authorization, one-to-one weapon and character relationship have been added

+++ Rpg/Services/CharacterService/CharacterService.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using Rpg.Data;
 using Rpg.Dto.CharacterDto;
 using Rpg.Models;
+using System.Security.Claims;
 
 namespace Rpg.Services.CharacterService
 {
@@ -10,46 +11,66 @@ namespace Rpg.Services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly ApplicationDbContext _dbContext;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, ApplicationDbContext dbContext)
+        public CharacterService(IMapper mapper, ApplicationDbContext dbContext, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _dbContext = dbContext;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.User = await _dbContext.Users.FirstOrDefaultAsync(c => c.Id == GetUserId());
+
             _dbContext.Characters.Add(character);
             await _dbContext.SaveChangesAsync(); // new id will be generated automatically
-            serviceResponse.Data = await _dbContext.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            serviceResponse.Data = await _dbContext.Characters
+                .Where(c => c.User.Id == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
             return serviceResponse;
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> DeleteCharacter(int id)
         {
-            var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
+            var response = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _dbContext.Characters.FirstAsync(c => c.Id == id);
-                _dbContext.Characters.Remove(character);
-                await _dbContext.SaveChangesAsync();
-                serviceResponse.Data =  _dbContext.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
-                return serviceResponse;
+                Character character = await _dbContext.Characters.FirstAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if(character != null)
+                {
+                    _dbContext.Characters.Remove(character);
+                    await _dbContext.SaveChangesAsync();
+                    response.Data = _dbContext.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    response.Status = false;
+                    response.Message = "Character not found";
+                }
+                return response;
             }
             catch (Exception ex)
             {
-                serviceResponse.Status = false;
-                serviceResponse.Message = ex.Message;
+                response.Status = false;
+                response.Message = ex.Message;
             }
-            return serviceResponse;
+            return response;
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _dbContext.Characters.ToListAsync();
+            var dbCharacters = await _dbContext.Characters
+                .Where(c => c.User.Id == GetUserId())
+                .ToListAsync();
             serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceResponse;
         }
@@ -57,35 +78,45 @@ namespace Rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetSingleCharacter(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _dbContext.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _dbContext.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
 
         public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdateCharacterDto updatedCharacter)
         {
-            var serviceResponse = new ServiceResponse<GetCharacterDto>();
+            var response = new ServiceResponse<GetCharacterDto>();
             try
             {
+                Character character = await _dbContext.Characters
+                    .Include(c => c.User)
+                    .FirstAsync(c => c.Id == updatedCharacter.Id);
 
-                Character character = await _dbContext.Characters.FirstAsync(c => c.Id == updatedCharacter.Id);
-                character.Name = updatedCharacter.Name;
-                character.Strength = updatedCharacter.Strength;
-                character.Defense = updatedCharacter.Defense;
-                character.HitPoints = updatedCharacter.HitPoints;
-                character.Intelligence = updatedCharacter.Intelligence;
-                character.Class = updatedCharacter.Class;
+                if (character.User.Id == GetUserId())
+                {
+                    character.Name = updatedCharacter.Name;
+                    character.Strength = updatedCharacter.Strength;
+                    character.Defense = updatedCharacter.Defense;
+                    character.HitPoints = updatedCharacter.HitPoints;
+                    character.Intelligence = updatedCharacter.Intelligence;
+                    character.Class = updatedCharacter.Class;
 
-                await _dbContext.SaveChangesAsync();
+                    await _dbContext.SaveChangesAsync();
 
-                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    response.Status = false;
+                    response.Message = "Character Not Found";
+                }
             }
             catch(Exception ex)
             {
-                serviceResponse.Status = false;
-                serviceResponse.Message = ex.Message;
+                response.Status = false;
+                response.Message = ex.Message;
             }
-            return serviceResponse;
+            return response;
         }
     }
 }
